Storage Proxy: Maintain metrics in response handler classes
There is a known problem introduced by 971711a, the problem
is that when some callbacks for a created response handler
are called from another scheduling groups an imbalance is
introduced to statistics which are now instantiated per
scheduling group. This can be easily solved by "remembering"
the stats of the scheduling group where the response handler
was created. This will mask out another problem where certain
operations are done outside of the "native" scheduling group.
This is a complex issue that needs to be addressed in separate
commit (see 971711a commit message).

Fixes #5797