table: add metrics for successful read latency
In order to distinguish between operations that succeeded and ones
that returned an error, a new histogram is introduced for successful
reads only. The new histogram is resilient to situations when
failing requests poison the "read latency histogram", showing
latencies higher than the timeout - so that it's still possible
to see what's the p99.9 latency for reads that actually succeeded.