reader_concurrency_semaphore: add linear backpressure
Once the reader concurrency queue is full, the old behavior was to
immediately refuse all surplus requests. Given very high client
concurrency, it's possible that Scylla would be very busy sending
refusals, causing the successful requests to have very high latency.
To make the issue less severe, requests are now refused with the
following heuristics, given max queue length Qmax:
 - requests before Qmax are not throttled
 - request Qmax+nth has n% chance to get throttled
 - when throttled, the request is refused, but it's first delayed
   by 10*n milliseconds.
For instance, if max queue is 100, request 142 has 42% chance
to get throttled, in which case it will be refused after being
delayed for 420ms.
The delays reduce the concurrency of refused requests, while still
allowing the successful ones to have decent latency.
The delay is capped to 1s in order to avoid ridiculous values.
The limit is based on local cassandra-stress tests, in which the delay
stabilized around 700ms.
Metrics in local tests showed that while the average latency for all
requests (including errors) raises, the latency of successful requests,
which is more meaningful, drops from >1s to sub-500ms. Also, the number
of failed requests drops, while the total number of served requests
remains the same.