gossiper: Run echo message handler on all shards
Echo message does not need to access gossip internal states.  Yes,
set_last_processed_message_at() is called but it is not really used. So
we can run it on all shards and avoid forwarding to shard zero.

This makes gossip marking node up more robust when shard zero is loaded.

Refs: #7197